requirements.txt
Django==4.2.7
django-allauth==0.57.0
django-axes==6.3.0
django-ckeditor==6.7.0
python-dotenv==1.0.0
geopy==2.4.1
django-select2==7.11.0
Pillow>=10.0.0
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-redis==5.4.0
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
watchdog==3.0.0 

